fix(authorizations): managing authorizations on a document now requires delete access
For the action to manage a document to show in menu items, the user now as to have delete access on the document similar to the action on a folder while also keeping previous validations. This change is to make sure the navigation to the authorizations page won't redirect to the home page due to it's access requirements. Grammar error fixes for authorization type 'share' also contained.

FD-8090