feature: Add ToAnInteger to ReferenceDependency
Add the method ToAnInteger to the class referenceDependency. As a side effect, oa bug was catch and corrected.

By default, in RecordImpl.correctValue, every instance of Number is converted to double. That was causing CastException when a metadata of type INTEGER was retreived from a record.

BREAKING CHANGE: Integer Type is corrected to int instead of double in RecordImpl.correctValue. Cast exception may happen